requirements.txt
python-telegram-bot[job-queue]==21.7
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.13.3
pytz==2024.2
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
greenlet==3.1.1
